add landing page, fix some bugs

+++ app/assets/javascripts/application.js
@@ -25,15 +25,16 @@ $(document).on('turbolinks:load',function(){
     // disable auto discover
     Dropzone.autoDiscover = false;
 
+    if($('.dropzone').length>0){
+        var dropzone = new Dropzone (".dropzone", {
+            maxFilesize: 256, // Set the maximum file size to 256 MB
+            paramName: "image[avatar]", // Rails expects the file upload to be something like model[field_name]
+            addRemoveLinks: false // Don't show remove links on dropzone itself.
+        });
 
-    var dropzone = new Dropzone (".dropzone", {
-        maxFilesize: 256, // Set the maximum file size to 256 MB
-        paramName: "image[avatar]", // Rails expects the file upload to be something like model[field_name]
-        addRemoveLinks: false // Don't show remove links on dropzone itself.
-    });
-
-    dropzone.on("success", function(file) {
-        this.removeFile(file)
-        $.getScript("/images")
-    })
+        dropzone.on("success", function(file) {
+            this.removeFile(file)
+            $.getScript("/images")
+        });
+    }
 });
